Changed my plugin manager to vim-plug
After some research realized that Vundle was no longer actively
maintained. There were some stiff alternatives, but went with vim-plug
as a lot of advanced vim users recommended it. It also has an easy
plugin management system and is very fast since processes run asyn.

dein.vim was also a note-worthy alternative, but it requires you to be
well-versed in vimscript, might consider it later.