Back to defaults!
Spent almost an entire day trying to make Vim a Vscode rip of with the
shiny features of code completion and what not. Realized that is not why
i was learning it in the first place, opted to only refer to docs and
not rely to heavily oon the code copletion feature.

Also YouCompleteMe and ConquerOfCode are **extremely** bloated in that
they slowed down my editor significantly. Hopefully wont have to come
back to them for now.

Also want to take some time off from the 'customizing my .vimrc'
loophole. Turns out I havent coded much this week, all because of the
vim frenzy.