Add keyd to dofiles
Keyd is a handy tool that allows remapping of my keys in the terminal.
It is however stored in the /etc/keyd directory and you have to link it
to enable tracking. Symlinking failed, and it ony traccks the path,
hence i decided to have a copy here, and my installation scriipt will
check all the stages and copy paste it there if succesful.

I wont be doing a lot of editing to it hence why i sis not even link the
two, it just remaps my escape key to caps and caps to escape. dassittt!!