Add some handy aliases to my .gitconfig
Always wondered why people kept the really short aliases like co to
checkout, and br for branch yet they werent that long to type oout.
Realized it gets kind on monotonous to type them over and over again and
this led me to adding some of the in my config.

This includes:
- che -> checkout
- br -> branch
- st -> status
- com -> commit
- visual -> to immediately open sublime merge(Mind blown!!)