feat(config/nvim): add toggle term to config
was getting pretty tired of having to open multiple terminals to
navigate runnign servers and such. installed toggle term, which allows
me to have a terminal within a terminal.

added some pretty cool keybinding to it to open the python REPL and also
the htop service immediately. The only downside is that i set it to
float mode that does not allow for multiple terminals.