chore(docs): updated the README.md
began my intial attempt of writing a script to reproduce the entire
config I i were to begin workign in another environment/os system.

Also had a lot of hiccups with wakatime as it was not monitoring my
activity any longer as such i removed it and installed it afresh again.
In my vim config, i returned the setting for scrollign  both up and down
in telescope to its defaults. I have gotten used to <C-n> and <C-p> for
upwards and downwards movements.